+++ changelog.md
@@ -6,10 +6,10 @@
 * Changed White Oak Leaves texture (now purple instead of orange)
 * White Oak no longer generate with leaf litters
 * White Oak now generate in Cherry Groves
+* White Oaks are now slightly less common in plains
 * Fallen White Oak no longer generate in Dark Forests
 * Prismarine blocks can now be used in conduit structure
 * Food crates and bread boxes can no longer be composted
-* Paper and Dried Kelp can no longer be used as fuel (vanilla behavior that was modified by the mod)
 * Added `beveled_glass_panes` block tag
 * Added `patterned_wool_stairs` & `patterned_wool_slabs` block/item tags
 * Removed now unused `nether_portal_frame_blocks` block tag